% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duration_month_chart.R
\name{duration_month_chart}
\alias{duration_month_chart}
\title{Display Episode Trend by Month}
\usage{
duration_month_chart()
}
\value{
ggplot2 figure
}
\description{
Display Episode Trend by Month
}
\examples{
duration_month_chart()
}
