% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duration_show_type_chart.R
\name{duration_show_type_chart}
\alias{duration_show_type_chart}
\title{Chart of Episode Duration by Show Type (i.e., Full vs. Bitesize)}
\usage{
duration_show_type_chart(episode_type = c("Full", "Bitesize"))
}
\arguments{
\item{episode_type}{character string. Full and/or Bitesize}
}
\value{
ggplot2 chart
}
\description{
Chart of Episode Duration by Show Type (i.e., Full vs. Bitesize)
}
\examples{
duration_show_type_chart(episode_type = c("Full"))
}
