% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duration_year_chart.R
\name{duration_year_chart}
\alias{duration_year_chart}
\title{Display Episode Trend by Year}
\usage{
duration_year_chart()
}
\value{
ggplot2 chart
}
\description{
Display Episode Trend by Year
}
\examples{
duration_year_chart()
}
