% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/episodes_table.R
\name{episodes_table}
\alias{episodes_table}
\title{Interactive Table Containing All Episodes and Audio Files}
\usage{
episodes_table()
}
\value{
DT table
}
\description{
Interactive Table Containing All Episodes and Audio Files
}
\examples{
episodes_table()
}
