% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paceypodcastR-package.R
\docType{package}
\name{paceypodcastR-package}
\alias{paceypodcastR}
\alias{paceypodcastR-package}
\title{paceypodcastR: Package for Scraping, Analyzing and Visualizing Pacey Performance Podcast Data}
\description{
This package was built as a tutorial on how to build packages in R for the Sportsmith R Course. It features functions for scraping all Pacey Performance Podcast episodes, text mining analysis, provides several vizualizations, and produces a shiny app.
}
\keyword{internal}
