% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timeline_chart.R
\name{timeline_chart}
\alias{timeline_chart}
\title{Timeline For Significant Pacey Performance Podcast Dates}
\usage{
timeline_chart()
}
\value{
ggplot2 figure
}
\description{
Timeline For Significant Pacey Performance Podcast Dates
}
\examples{
timeline_chart()
}
