% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wordcloud_figure.R
\name{wordcloud_figure}
\alias{wordcloud_figure}
\title{Pacey Performance Podcast Wordcloud}
\usage{
wordcloud_figure()
}
\value{
wordcloud figure
}
\description{
Pacey Performance Podcast Wordcloud
}
\examples{
wordcloud_figure()
}
